requirements.txt
numpy~=2.3.3
pillow~=11.3.0
pip~=25.2
attrs~=25.4.0
protobuf~=6.32.1
tornado~=6.5.2
Jinja2~=3.1.6
pytz~=2025.2
toml~=0.10.2
charset-normalizer~=3.4.3
setuptools~=65.5.0
numba~=0.62.1
regex~=2025.9.18
smmap~=5.0.2
sympy~=1.14.0
mpmath~=1.3.0
decorator~=4.4.2
click~=8.3.0
pyarrow~=21.0.0
pandas~=2.3.3
packaging~=25.0
tzdata~=2025.2
scipy~=1.16.2
colorama~=0.4.6
urllib3~=2.5.0
networkx~=3.5
tenacity~=9.1.2
watchdog~=6.0.0
streamlit~=1.50.0
openai-whisper~=20250625
moviepy~=2.0.0.dev2